+++ email_archiver/__init__.py
@@ -1,2 +1,2 @@
 """Email-to-EML Secure Archiver"""
-__version__ = "0.5.0"
+__version__ = "0.6.0"